requirements.txt

openai>=0.27.0
streamlit==1.40.2
phidata==2.7.3
Pillow==10.0.0
duckduckgo-search==6.4.1
google-generativeai==0.8.3
